labirinto/requirements.txt
matplotlib==3.7.5
numpy==1.24.4
pygame==2.6.1
pyopengl==3.1.10